+++ include/PVX_String.h
@@ -13,7 +13,7 @@ namespace PVX {
 	namespace String {
 		inline void OnSplit(const std::string& Text, const std::string& Separator, std::function<void(const std::string&)> clb) {
 			size_t ssz = Separator.size(), last = 0;
-			int start = 0;
+			long long start = 0;
 			while (-1 != (start = Text.find(Separator, start))) {
 				if (((unsigned int)start) >= last)
 					clb(Text.substr(last, start - last));
@@ -26,7 +26,7 @@ namespace PVX {
 
 		inline void OnSplit(const std::wstring& Text, const std::wstring& Separator, std::function<void(const std::wstring&)> clb) {
 			size_t ssz = Separator.size(), last = 0;
-			int start = 0;
+			long long start = 0;
 			while (-1 != (start = Text.find(Separator, start))) {
 				if (((unsigned int)start) >= last)
 					clb(Text.substr(last, start - last));
@@ -108,7 +108,7 @@ namespace PVX {
 	namespace StringView {
 		inline void OnSplit(const std::string_view& Text, const std::string_view& Separator, std::function<void(const std::string_view&)> clb) {
 			size_t ssz = Separator.size(), last = 0;
-			int start = 0;
+			long long start = 0;
 			while (-1 != (start = Text.find(Separator, start))) {
 				if (((unsigned int)start) >= last)
 					clb(Text.substr(last, start - last));
@@ -121,7 +121,7 @@ namespace PVX {
 
 		inline void OnSplit(const std::wstring_view& Text, const std::wstring_view& Separator, std::function<void(const std::wstring_view&)> clb) {
 			size_t ssz = Separator.size(), last = 0;
-			int start = 0;
+			long long start = 0;
 			while (-1 != (start = Text.find(Separator, start))) {
 				if (((unsigned int)start) >= last)
 					clb(Text.substr(last, start - last));
